Moved special attack menu handling into Player instead of Battler, since enemies cannot select from visible special menus
